# support SIMD for larger symmetric Bit-exact 16U gaussian blur