## perf test
env:
Parallel framework: pthreads (nthreads=8)
CPU features: SSE SSE2 SSE3 *SSE4.1 *SSE4.2 *FP16 *AVX *AVX2 *AVX512-SKX?

<cut/>

Measured performance of `GaussianBlur(src, dst, Size(7,7), 0, 0, btype)`.
Conventional perftests (3x3 and 5x5) are not improved because these are already optimized in 698b2bf.

| name | image size, type, btype | mean (usec) in 3.4 | mean (usec) in this PR |
|---|---|---|---|
gaussianBlur7x7/0| 	(640x480, 16UC1, BORDER_REPLICATE) | 358 |100
gaussianBlur7x7/1 |	(640x480, 16UC1, BORDER_CONSTANT)| 351|97
gaussianBlur7x7/2 |	(640x480, 16UC1, BORDER_REFLECT) 	| 605|100
gaussianBlur7x7/3 |	(640x480, 16UC1, BORDER_REFLECT101)| 694|100
gaussianBlur7x7/4 |	(1280x720, 16UC1, BORDER_REPLICATE) |2081|343
gaussianBlur7x7/5 |	(1280x720, 16UC1, BORDER_CONSTANT)| 1219|489
gaussianBlur7x7/6 |	(1280x720, 16UC1, BORDER_REFLECT) 	|1221|549
gaussianBlur7x7/7 |	(1280x720, 16UC1, BORDER_REFLECT101)|1453|422

### Pull Request Readiness Checklist

See details at https://github.com/opencv/opencv/wiki/How_to_contribute#making-a-good-pull-request

- [x] I agree to contribute to the project under Apache 2 License.
- [x] To the best of my knowledge, the proposed patch is not based on a code under GPL or other license that is incompatible with OpenCV
- [x] The PR is proposed to proper branch
- [x] There is reference to original bug report and related work
- [x] There is accuracy test, performance test and test data in opencv_extra repository, if applicable
      Patch to opencv_extra has the same branch name.
- [x] The feature is well documented and sample code can be built with the project CMake

```
force_builders=Linux AVX2,Custom
buildworker:Custom=linux-3
build_image:Custom=ubuntu:18.04
CPU_BASELINE:Custom=AVX512_SKX
disable_ipp=ON
```