@terfendail 2020-12-02
@Yosshi999 Could you please extend accuracy tests to 16U gaussian blur?