@Yosshi999 2020-12-03
@terfendail 
This one https://github.com/opencv/opencv/blob/3.4/modules/imgproc/test/test_smooth_bitexact.cpp#L184 is not enough?