@alalek 2020-12-10
Results from `DISABLED_FULL/OCL_GaussianBlurFixture.GaussianBlur*` perf tests.

Configuration: i5-6600, 1 thread, no IPP
Parameters: `--gtest_also_run_disabled_tests --gtest_filter=*FULL*auss* --perf_threads=1`

|Name of Test|base|patch|speedup|
|---|:-:|:-:|:-:|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 3)|7.498|7.071|1.06|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 5)|12.050|13.087|0.92|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 7)|24.956|4.365|5.72|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 9)|31.764|5.215|6.09|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 11)|38.540|6.356|6.06|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 13)|44.926|7.654|5.87|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 15)|51.635|8.848|5.84|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 17)|58.032|9.443|6.15|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 19)|64.759|10.835|5.98|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 21)|71.598|12.189|5.87|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 23)|78.125|13.729|5.69|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 25)|85.296|15.249|5.59|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 27)|92.261|15.528|5.94|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC1, 29)|99.694|16.858|5.91|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 3)|15.323|14.170|1.08|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 5)|23.935|26.356|0.91|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 7)|49.970|8.383|5.96|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 9)|63.157|10.710|5.90|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 11)|76.411|12.999|5.88|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 13)|89.464|15.747|5.68|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 15)|103.414|16.913|6.11|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 17)|116.021|19.377|5.99|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 19)|129.263|22.736|5.69|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 21)|143.478|25.887|5.54|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 23)|156.401|26.601|5.88|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 25)|170.235|29.110|5.85|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 27)|184.335|32.559|5.66|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC2, 29)|197.766|35.903|5.51|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 3)|23.365|21.230|1.10|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 5)|35.730|39.332|0.91|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 7)|75.842|13.147|5.77|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 9)|95.262|16.609|5.74|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 11)|115.406|18.734|6.16|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 13)|135.004|22.408|6.02|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 15)|155.203|26.811|5.79|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 17)|175.163|30.711|5.70|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 19)|195.481|32.315|6.05|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 21)|215.616|35.873|6.01|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 23)|235.929|41.340|5.71|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 25)|255.522|45.972|5.56|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 27)|276.735|46.550|5.94|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC3, 29)|297.379|50.300|5.91|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 3)|31.000|28.539|1.09|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 5)|47.669|52.243|0.91|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 7)|100.822|16.707|6.03|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 9)|126.625|21.203|5.97|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 11)|153.166|26.250|5.83|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 13)|180.047|31.758|5.67|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 15)|207.392|34.135|6.08|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 17)|237.175|40.819|5.81|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 19)|270.267|48.091|5.62|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 21)|302.194|53.799|5.62|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 23)|335.328|61.909|5.42|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 25)|368.538|66.779|5.52|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 27)|401.667|70.240|5.72|
|GaussianBlur::DISABLED_FULL/OCL_GaussianBlurFixture::(1920x1080, 16UC4, 29)|427.746|77.757|5.50|