@alalek 2020-12-10
Any thoughts about 8-9% regressions with 5x5 kernel?