@Yosshi999 2020-12-11
> Any thoughts about 8-9% regressions with 5x5 kernel?

I have no idea. But implementing SIMD version of [5Nabcba](https://github.com/opencv/opencv/blob/6b75e4ddd697442b42cb6206bf8735e06336221b/modules/imgproc/src/smooth.simd.hpp#L757) for `ufixedpoint32` may improve its speed.